Output module files for all entrypoints (#2220)
Previously _almost_ all entrypoint libraries were considered the primary
source for their module. The only ones that were not considered primary
were those that were in an import cycle with another entrypoint. After
https://github.com/dart-lang/build/pull/2125 this is no longer the case,
the primary source may be some other library. The change breaks the
discovery of modules for entrypoints that contains a `main` from the
compiler packages - it was already broken for the very rare case of
import cycles from an entrypoint.

Detect the case where we won't output a module for an entrypoint and
output one anyway. This will break any behavior which is depending on
the `.module` file _only_ corresponding to the primary source. Primary
source is only a concept used for discovery so this should be safe - it
only makes modules easier to discover.