Refactor resolver crawling (#2244)
Fixes https://github.com/dart-lang/sdk/issues/36773

Previously the crawling was intermingled with handling changes and we
had a single use implementation of `FutureGroup` to capture all the work
happening. Refactor to use `crawlAsync` from `package:graphs` for the
crawling portion, and separate out the crawling from notifying the driver
that the contents have been changed. By the time `driver.changeFile` is
called the `resourceProvider` should have updated content for imported
libraries.

Since the `_cache*` maps are used to know what specifically needs to be
changed in the `resourceProvider`, the updates to both places needs to
happen synchronously. 

Update the invalidation of dependency cache to store asset digests
instead of contents.