Allow resource value reuse when disposed
The intent with resources was to be able to do things like cache
expensive computations across builds, with the `dispose` hook helping to
keep things as hermetic as possible. Throwing away the entire instance
defeats the purpose.

Only discard the instance if there was no dispose callback
available.

Expand the class level docs to make the usage and intent more clear.