Improve failure for dart2js args
Fixes #2353

Since the yaml parsing may give us something other than a `String` we
can use `toString` instead of a cast which leaves us as agnostic towards
the args that dart2js allows as possible. With this change the user will
see an output from dart2js usage with `Error: Unknown option '-3'.`.
This is still a little confusing because the `-03` turned into `-3`.