Add an option to skip the platform check for specific packages (#2420)
Flutter web must allow some `dart:io` import in `package:flutter`, but
adding `dart:io` to the `DartPlatform` causes the module crawling to
pick the imports in some cases - we want the `DartPlatform` to match the
`libraries.json` that will be used by DDC and kernel in terms of which
libraries are supported, but to avoid the failures for bad imports.

The kernel and DDC workers have always allowed imports to missing
`dart:` libraries with stubbed implementations that will throw if they are
used at runtime. The restrictions on imports are implemented in the
build systems instead of the compilers. This argument allows loosening
the restriction in specific flutter packages while keeping the handrails for
user code.