ensure sdk dependencies are built as a part of ddc bootstrapping (#2445)
Specifically this allows simple build filters to build any dart web app - you can just ask for the `*.dart.js` file and the entire app will be built.

This works because the SDK files are now inputs to the bootstrapping action. This also has other nice side effects of ensuring that the required SDK files do in fact exist where we expect them.