Create a custom summary from the users sdk (#2467)
Fixes https://github.com/dart-lang/sdk/issues/38499.
Fixes https://github.com/dart-lang/build/issues/2451.

Lazily generates an sdk summary from the users SDK files and caches it under `.dart_tool/build_resolvers`. On my machine this adds ~3 seconds on first use.

The summary is invalidated based on the `Platform.version` which is cached in a file next to the summary.

A custom summary can also be provided (for non-pub package scenarios). I am going to validate this can be rolled successfully internally before landing.

Note that this has one nice property of allowing build_runner and build_test to share the summary file, which is one advantage over implementing this with a Builder.