Give a more consistent ordering for builders
The strongly connected components algorithm doesn't make any guarantees
about the ordering of values that don't depend on eachother, when there
are multiple valid topological orderings it doesn't make any guarantees
about which it will use. We can count on it being consistent though, it
is based on the order of the children it sees from `edges` for each
node.

The ordering of `builders` when we look for the topological sort may
depend on both the ordering of the topological sort of the package
graph, as well as the order it iterates keys within the builder
definitions from the parsed `build.yaml`. Since these aren't guaranteed
we also don't have a stable ordering of the builders to start with. The
order of the edges is based on the order we see them here.

Since we don't care about _which_ arbitrary ordering we get, it should
be sufficient to pick some to start with. We can sort by builder keys
which are guaranteed to be unique.