Enforce ordering between SDK JS copier and DDC
The DDC entrypoint builder will ensure that the SDK resources are
available with `buildStep.canRead` and so they will be required to exist
in an earlier phase. This is going to be true for almost all packages
anyway since the `sdk_js_copy` builder only applies to
`build_web_compilers` will will almost always be non-root and not in a
dependency cycle with the root. For the dependencies of
`build_web_compilers` we can make it a guarantee.