add windows tests where they pass, fix some bugs (#2488)
Will remove appveyor config later - have to get the webhook deleted first

This adds windows bots for all our unit and integration tests with a few skips here and there.

The major exception are tests in the build_runner package which were having a lot of problems, and all the meaningful build_daemon tests. Both were primarily caused by issues with cleaning up temp directories after the test but the build daemon is still running in that directory.

I also had to do a few minor fixes in the scratch_space and build_test packages so this releases new versions of each of those.