Ensure all help is handled in outer binary
Fixes #2498, fixes #2499

Previously we'd only handle help in the outer binary if it was invoked
specifically as the `help` command, or if no command was given. If we
passed `-h` as a sub argument to another command we wouldn't recognize
it and would still generate the build script. We also always printed the
top level usage rather than the usage that would be printed by the real
help command, which may be specific to a sub command.

- Dig 2 levels deep for the `-h` or `--help` flags. This way it can be
  passed as either a top level flag, or a flag for the specific sub
  command.
- When printing usage for help specifically, use the command runner
  since it already implements figuring out whether the print outer usage
  or usage for a sub command.
- Handle the missing command as it's own case after we know we didn't
  handle help. Set the exit code to `ExitCode.usage` for this case, and
  print the top level usage.

This pattern won't work if we start adding nested commands more deeply
within other commands since we only pick up the `-h` from the top level
or 1 command deep.