Invalidate cached modules by digest
Within the build that writes a new value we don't have the digest so at
the end of the build the value will be cleared away and must be read
from the file at least once.
As soon as a value is read from a file we also track the digest and only
invalidate the deserialized value when the digest has changed.