Improve error message for a bad hostname
Closes #1941

When using the `--hostname` argument with a name that isn't supported by
the OS the `SocketException` would have a null `address` and `port`
field. In this case we'd get a stack trace trying to emit an error which
makes it look like a bug in `build_runner` rather than a usage error.

Since we don't know all the conditions that lead to a `SocketException`
the message is somewhat vague.