Fix strict-inference hints in build_runner
Add return types on all methods, including `main`. Mostly the missing
types were `void`.

Add argument types on all arguments where they can't be inferred. This
is a clear improvement in a few places where we otherwise had dynamic
calls in the body. In some places where we are ignoring the argument
change the argument type to `void` to make it clear it won't be used.
Remove one unnecessary optional argument since it wasn't required to fit
the interface we were implementing.

Add a `void` return type to function types where we don't use the
return. In some places switch to the newer function type syntax since it
is more clear.