Add a warning about DDC deprecation
Closes #2564

Add a package version that is only picked up by recent SDK releases that
warns that DDC is no longer supported and directs the user to upgrade to
newer releases of the `build_web_compilers` package.

Fail the build during Builder instantiation to avoid doing work that
will be thrown away and spamming the output with errors for every
module.