Fix resolver behavior when encountering non-existing asset ids (#2618)
I follow the advice from https://github.com/dart-lang/build/issues/2389#issuecomment-582595934. Basically, we're not supposed to return a null source if the uri could refer to a valid asset id. However, we used to return null if we haven't seen the asset id yet (e.g. it doesn't exist). Now, we just return the source pointing to the file where a non-existing asset id would be. If it gets created in a later build step, the analyzer will recognize this and handle it correctly.

With manual testing, I can verify that this indeed fixes #2389. Also contains, and thus closes #2392.