Remove dependencies on package_resolver
Towards #2640

Remove the only non-test dependency from `package:build_resolvers`.
Replace with directly using `Isolate.resolvePackageUri` which
is what `PackageResolver.current` does for the API we are using.

Move some other dependencies to `dev_dependencies` since the package is
only used in tests. We will want to follow up with removing it
completely.