Depend on package language versions instead of entire package_graph.json (#2659)
Fixes https://github.com/dart-lang/build/issues/2645

This adds a map of package name to language version to the asset graph, and checks that during deserialization. If it changes we throw away the graph and restart.

Note that this is a breaking change in build_runner_core, as some factory methods are now static methods that return a Future, and `PackageNode` now also has a language version.

This also reverts some of the previous logic and moves the check from build_script_updates.dart to build_definition.dart.