use addTearDown instead of tearDown to fix race condition (#2665)
Fixes flaky build_daemon tests.

The `test_descriptor` package uses `addTearDown`, which always runs before `tearDown`. This makes it attempt to clean up the temp directory while the daemon is still running. As it deletes files the daemon tries to recreate them and if it does that before the recursive delete is finished then the delete fails.

By using `addTearDown` ourselves - and after creating the test descriptor dir, we ensure the daemon process is cleaned up _first_.