Reduce the chances of different build phase order (#2674)
Fixes #2645

Sort the list of packages we get from `PackageConfig`. This won't fully
guarantee consistency since we later have behavior that doesn't make
ordering guarantees, but in practice with a consistent package order to
start with we _should_ get a consistent build phase order.

Also sort dependencies from pubspecs so that the "children" calls for
the package graph use a consistent order.

This won't _guarantee_ ordering of build phases, but they should be
consistent in the vast majority of cases, and when they are not it won't
introduce a correctness bug, only a missed caching optimization.