Remove uses of the term whitelist (#2716)
Rename to "sources" where applicable. The term "default" already conveys
the meaning we need, the "whitelist" term isn't useful since it's not
used as a filter, only as a default set of sources.

`defaultRootPackageWhitelist` was not used outside the package anyway,
so remove it from the public API. Also do an audit of other public
members from `constants.dart` that aren't used outside of the package
and add an explicit `show` to the export.