MockBuilder: Do not be concerned with method modifiers.
An overriding method and an overridden method need not have matching modifiers. For example, an async method can be overridden by a synchronous method; the modifiers are not part of the signature.

Introduce support for returning a simple dummy Iterable (an empty List) and a simple dummy Stream (an empty Stream).

Improve existing tests and introduce new tests.

PiperOrigin-RevId: 315357440