Run a silent help before generate-build-script (#2726)
Fixes #2725

The first `pub run build_runner` after a `pub get` will print some lines
about "Precompiling" the target. This throws off the VM when it's passed
as the argument to the `dart` command. Prefix with a `pub run
build_runner help` but send the output to `/dev/null` to suppress it.
This is the command that will print the "Precompiling" message, then the
next one will print only the path to the script.