Remove the export of constants from _test_common
This export is overly broad and masks the places where a test _should_
be getting an import from the public library instead of `src/`. This
makes it more clear which tests are using the internal details.

If this had been the pattern before the issue in #2772 _may_ have been
more obvious. Its a better pattern to avoid this in either case.