Use : over | for builder keys
We have had the ability to parse `:` for a while but did not update
documentation to use it since users on older versions of
`package:build_config` would have issues. Since it has been out for a
long time we expect most users to be on the latest. Update docs to use
`:` instead of `|` for the `build_web_compilers:entrypoint` builder.
There may still be other uses for other builders in docs.

The important update is that the arguments passed on windows on travis
use the `:` character since `|` causes problems with argument passing.