Add a test demonstrating potential bad analysis
We recently saw trouble with imports that appeared to be missing due to
the `parseFile` API causing unexpected reads and caching the state of
the file as missing before we had made it available. The usage of
`parseFile` will become safe, but when a file is generated there can be
other situations where the analyzer reads it as missing and caches that
fact.

Add a test with a series of builders demonstrating a specific series of
interactions that lead to this problem.

From the standpoint of the analyzer, we need to perform two steps in
order for the AnalysisDriver to see the correct content for a file after
it has previously tried to read it and it was missing:
1 - MemoryResourceManager.newFile
2 - AnalysisDriver.changeFile

In the case where the analyzer had never cached any state for that file,
we can get by with just `newFile`. We do currently guarantee that
`newFile` is called in time for analysis, but we don't guarantee that
`changeFile` has been called. In this example we have the following
interactions.

- g1 builder: reads a file which imports the out it has not yet written.
  Puts the analayzer into the cached bad state.
- root builder: (next non-optional builder) uses the resolver and starts
  a transitive crawl.
  - This crawl sees `.g1.dart` and calls `newFile`. It also caches the
    digest.
  - This crawl asks for `.slow.dart` and`.g2.dart`. This triggers a
    lazy build in both.
- The builder for `.slow.dart` won't complete, therefore the
  `crawlAsync` being done for the resolver intended for the "root"
  builder won't complete, therefore no `changeFile` are called.
- The builder for `.g2.dart` tries to use the resolver. During it's
  transitive dependency crawl it found that the digest for `.g1.dart`
  was cached, so it skips this asset for `changeFile`. It sees "null"
  for the annotation value, instead of what it would get if the
  `.g1.dart` import were correctly visible.