Fix handling of changeFile
Remove the concept of deciding concretely whether to call `changeFile`
during the crawl. Instead track the asset paths that currently need to
be updated in the `AnalysisDriver` in a separate set. The first build
step to see the asset, or changed asset, during a transitive crawl will
record it as needing an update, and the first build step to finish a
crawl that visited that asset will call `changeFile`. Often these will
be the same build step.

- Add a `_needsChangeFile` set. The extra memory used should scale with
  ongoing work, not with the total size of the project.
- Simplify the `_AssetState` class.
- Add an extra check against the cached digests after reading the file
  content. This prevents duplicate work of `_parseDirectives` in the
  case where multiple build steps saw an uncached digest and start to
  await the `readAsString` call.