Fix handling of changeFile for the analysis driver (#2866)
Fixes an issue where we may fail to notify the `AnalysisDriver` of a
newly available asset that it has previously read as missing before we
provided it. This could happen in some versions of the analyzer package
when calling `parseFile`, and it can happen in all versions when we use
the resolver with an import to an asset that is not generated yet, so
not visible to the build step.

See https://github.com/dart-lang/sdk/issues/43744

From the standpoint of the analyzer, we need to perform two steps in
order for the AnalysisDriver to see the correct content for a file after
it has previously tried to read it and it was missing:
1 - MemoryResourceManager.newFile
2 - AnalysisDriver.changeFile

In the case where the analyzer had never cached any state for that file,
we can get by with just `newFile`. We do currently guarantee that
`newFile` is called in time for analysis, but we don't guarantee that
`changeFile` has been called.

Remove the concept of deciding concretely whether to call `changeFile`
during the crawl. Instead track the asset paths that currently need to
be updated in the `AnalysisDriver` in a separate set. The first build
step to see the asset, or changed asset, during a transitive crawl will
record it as needing an update, and the first build step to finish a
crawl that visited that asset will call `changeFile`. Often these will
be the same build step.

Resolver changes to fix:
- Add a `_needsChangeFile` set. The extra memory used should scale with
  ongoing work, not with the total size of the project.
- Simplify the `_AssetState` class.
- After any crawl, find the intersection between the transitive
  dependencies crawled, and the assets that need an update, and call
  `changeFile` for each.

Also, as a separate optimization: Add an extra check against the cached
digests after reading the file content. This prevents duplicate work of
`_parseDirectives` in the case where multiple build steps saw an
uncached digest and start to await the `readAsString` call.

Add a test with builders forcing a specific series of interactions that
demonstrates this problem:

- g1 builder: reads a file which imports the out it has not yet written.
  Puts the analyzer into the cached bad state.
- root builder: (next non-optional builder) uses the resolver and starts
  a transitive crawl.
  - This crawl sees `.g1.dart` and calls `newFile`. It also caches the
    digest.
  - This crawl asks for `.slow.dart` and`.g2.dart`. This triggers a
    lazy build in both.
- The builder for `.slow.dart` won't complete, therefore the
  `crawlAsync` being done for the resolver intended for the "root"
  builder won't complete, therefore no `changeFile` are called.
- The builder for `.g2.dart` tries to use the resolver. During it's
  transitive dependency crawl it found that the digest for `.g1.dart`
  was cached, so it skips this asset for `changeFile`. It sees "null"
  for the annotation value, instead of what it would get if the
  `.g1.dart` import were correctly visible.

Remove the `invariant_booleans` lint rule for false positives.