Add support for directly debugging browser test
Closes #2749

In the generated `debug.html` file append the query parameter
`directRun=true`, and check for this parameter in the generated
`.browser_test.dart` files to determine whether to bootstrap through the
test runner, or to run the test directly.

Use the `.browser_test.dart.js` file as the JS script to load from the
`.debug.html` file.

Refactor a `p.url.join` to `Uri(pathSegments)`. This also URL encodes
the segments if necessary, and allows adding the query parameter.