Expand readable assets for non-root packages, make them configurable (#2851)
Instead of throwing an `InvalidInputException` for non-lib, non-root packages, we now allow the following files by default:
- `lib/**`
- `bin/**`
- `LICENSE`
- `README*`
- `pubspec.yaml`

Each package can expose additional globs with the `additional_public_assets` build option. I opted to not support exluding assets because
- excluding `lib/` might lead to unexpected behavior
- that package's default target includes `additional_public_assets` by default. If there were two `excludes`, it's not clear in which priority they should interfere with each other

In the root package, all files are still allowed of course. I moved the check to throw an `InvalidInputException` down to the `SingleStepReader`. Computing the globs of available sources per package now happens in the `TargetGraph` since that's where the build config is read. That feels a bit unclean since the _target_ graph now decides which assets are visible for a _package_, please let me know if I should move that logic to the `PackageGraph` or elsewhere.

Closes #2819. I also wrote [an example](https://github.com/simolus3/build_license_collector/blob/main/lib/src/crawler.dart) to crawl licenses across dependencies.