Codegen API: Override inherited methods and fields.
Users expect to be able to stub and verify inherited methods and fields. This work supports that.

The tricky parts were making sure we override using substituted type parameters. When writing a MockFoo class for `class Foo extends FooBase<int>`, we need to substitute `int` everywhere `T` is found in a signature.

PiperOrigin-RevId: 343594212