Expose a new api to get the AstNode for an Element (#2947)
The api is `Future<AstNode> astNodeFor(Element element, {bool resolve: false})`, and is exposed on `Resolver`.

This api can be used avoid using the AnalysisSession directly, which is unsafe to do in Builder code and results in `InconsistentAnalysisException`s. Internally we can avoid this issue due to having synchronous access to the current analysis session.

Fixes https://github.com/dart-lang/build/issues/2946.