Bump min dependency on build
This is an indirect way to limit which versions of `build_runner` will
be compatible with this package. This change relies on #2955 to work
correctly. The `build` constraint won't directly solve that, however it
will limit the scope of which `build_runner` versions are compatible to
just one which does not include the fix -`1.10.10`, which should always
be upgradeable to `1.10.11`.