Scrub unnecessary term from changelogs
The code was already updated with the changelog left as is. While it's
not entirely clear that these removals are critical - it's easier to
just update it than to try to make the detection tools less sensitive.

Removes some specific references to code members that were removed in a
changelog entry that is old enough that the value of keeping the
specifics in the changelog is low. Rewrites other uses in prose since it
doesn't matter that they don't specifically match the names used in the
code at the time the changelog was written.