Migrate user-facing apis to null-safety (#2996)
Migrate packages typically imported by build authors to null-safety (part 1 /3). Closes #2920

In this PR, I migrate

- the `build` package
- the `build_test` package (or rather, the parts of that package that don't import `build_resolvers`)
- the `scratch_space` package

After the `graphs` package is migrated, I'll open another PR to migrate the `build_resolvers` package. When that one is done too, we can start to opt the tests of `build` and `build_test` into null-safety and publish, thus enabling build authors to write and test their builders with strong null-safety. 

Migrating the rest of the build system is not a goal for the near future, so we won't run builds with strong null-safety.