Migrate uses of followedBy to collection literals (#3130)
In the past I've looked for other common patterns that look nicer as
literals and on a hunch I wanted to check which uses of `followedBy` are
more readable as subsequent elements in a collection. In these cases I
think it is an improvement.

Most migrated cases either were immediately gathered into a collection
to `.toList` or `.toSet`. In a few places where I trust that the
collection will be fairly small and always iterated I also switched from
using a lazy iterable into an eager collection.

Some uses of `followedBy` remain. Those cases are either where an eager
collection could be non-performant, or cases where the iterable chains
subsequent `.map` or other calls that would make the collection literal
more noisy.