Pass error and stack trace for request failure log
Correct the `_logger.severe` call to pass the error (instead of a
hardcode `true`) and stack trace to the log record instead of manually
formatting a full message.

Refactor the logging middleware to async/await for better readability,
and to allow using `rethrow` which does not to ignore a lint about only
throwing subtypes of Error or Exception.

Rename `_getMessage` to `_requestLabel` to be more descriptive about the
message it creates.