Avoid folding frames for build system bugs
The intent of the `.foldFrames` call is to reduce the noise for stack
traces when an exception is thrown inside a builder implementation. The
folding makes internal bugs harder to diagnose though, because it can
mask the source of exceptions that are thrown within `build_runner`
code.

Check if the lowest frame of the trace is within one of the build
system implementation packages, and if it is do no folding.

Add `build_runner_core` to the list of packages treated as
implementation of the build system.