Use correct class when reviving from a static field (dart-lang/source_gen#546)
Fixes dart-lang/source_gen#533

Previously when searching for a static field matching a constant object
we would use the class name from the type of the object always instead
of the class name in which we found the field. This worked in many cases
because constant static fields intended for use in other libraries are
more often defined on the same class as their type.