Update DDC entrypoint outputs to support multiple
Currently, serving a page that loads more than one `.dart.js` entrypoint
(when they are compiled with DDC) results in only the last one actually
executing. This is due to the way the entrypoint JS loads requirejs and
queues up the loading of its bootstrap file. Previously, it used the
`data-main` attribute on the require.js `<script>` element, but that
only works if there's only one main to call. In the case of multiple
entrypoints, each one effectively overwrites the previous `data-main`.

This commit changes the way the bootstrap module is loaded and
executed to a way that is compatible with multiple DDC entrypoints on
the same page. This is accomplished by generating an additional
`.bootstrap.require.js` output from the `build_web_compilers|entrypoint`
builder that will be responsible for calling `require([...])` with the
bootstrap module name, and then updating the entrypoint JS output to
load this new asset after loading `require.js`.