Label assets without .dart as not Dart libraries (#3112)
The old behavior was closer to `isNotDartPartFile` because it treated
any file without the content `part of` as if it was a Dart library. Add
an extension check to remove these false positives.

This may introduce some false negatives if any projects relied on
resolving Dart code in files with other extensions.