Allow resource value reuse when disposed (#3207)
The intent with resources was to be able to do things like cache
expensive computations across builds, with the `dispose` hook helping to
keep things as hermetic as possible. Throwing away the entire instance
defeats the purpose.

Only discard the instance if there was no dispose callback
available.

Change the handling of disposing resources that haven't been used to an
`assert`. The prior fallback behavior was never hit, and we don't want to
 silently create resources that aren't needed solely to dispose them if we
 introduce a bug that would cause it.

Expand the class level docs to make the usage and intent more clear.