Use raw string literals in more places when generating the build script (#3214)
When generating the build script, some string literals like the builder's name are generated as raw strings already. However, strings that might be part of a builder's options are not escaped.

For example, a build script generated for the following builder options would fail to run:

```yaml
options:
  args: "-shared -o $output $input"
```

With this PR, we use raw string literals in more places. I have updated the golden file as a test, let me know if we need more tests for this.