Add overrideUnsupportedMembers option to MockSpec
In migrating much mockito code to code generation, I've found that in the vast majority of fallbackGenerators use cases, a fallback generator is used exclusively to satisfy the type contract, and is never used at runtime, because the method is never used in the test into which it is generated.

In these cases it seems silly to require users to carefully type out this fallback function and choose a body for it (throw this? throw that? try to account for return types and return things?).

Also fix an issue with fallback generators and getters.

PiperOrigin-RevId: 427214894