await a missed test failure, fix broken test (#3245)
This test should be failing but the error is lost because of the way the
async failure interacts with the `runBuilder` error handling zone.
https://github.com/dart-lang/test/issues/1670

Use `await expectLater` instead of `await expect` to ensure the error is
not lost.

Change the setup for the test to include two imports instead of 1 import
and a local class definition. Move the usage to an annotation so that it
is easy to find from the library using it.