Return nullable Rejection from softCheck
I hope this allows more expressive (but still sensible) output in some
cases where a soft check is expected to match against a specific
element. In other cases where it may not be expected to match, or allows
multiple elements, there isn't a single specific rejection that is
interesting.

Pass both the full formatted message and the nullable rejection to the
failure callback. It doesn't make sense to have the failure callback do
the formatting because it is tightly couple with `_TestContext`