Use : as builder key separator in docs
Closes #3239

The character was originally `|` but was changed to `:` because the
former is a special character in windows terminals. Some docs where
updated with that change, but some were missed.

Use `:` over `|` consistently in docs, `build.yaml`, and command line
arguments in tests.

Fix a utility which shortens redundant builder names to use the new
separator. This utility had not been working.