Use explicit generics for collections in build script (#3301)
Towards #3300

Avoid inference to improve consistency between getting configuration in
from default options or from user provided options which are a
`YamlList` or `YamlMap` and never have a generic other than `dynamic`.
The generic is the most user visible difference, the type of the
concrete collection is not as likely to cause problems. Use `String` keys
for maps since the `BuilderOptions` class requires it for the outer
collection.

Avoid `toExpression` for the `appliesBuilders` key to keep using
inference for that list. Now `toExpression` is only for the default
options. Remove now unused support for const in `toExpression`.