Remove tests that attempt to modify files
Close #3202

Calling `MemoryResourceProvider.modifyFile` is not sufficient for the
analyzer to respond to the changes, `AnalysisDriver.changeFile` is the
important call, but we don't have access to that member from the tests.
The real world behavior that would lead to this scenario requires more
complex interactions we can do with `resolveSources`.

Remove these tests entirely since they are not testing what they purport
and are not useful as they stand.